Split reliable agent browser journeys
